Extend support for Argent signatures with guardian signer (#25)
A signature format of `5` segments refer to Argent accounts with a
single signer. Accounts that activate a "Smart Account" will have a
second signer "Guardian" with the `9` segments.

This change will also allow a user with a regular account who then signs
up for smart account to continue to use Paradex.