Check if wallet generates deterministic signatures (#32)
This pull request includes changes to enhance the deterministic signing
process and improve error handling in the `src/account.ts` file, as well
as a minor update to the `README.md` file.

Enhancements to deterministic signing:

*
[`src/account.ts`](diffhunk://#diff-ae151444b91b7e5e0a6aa753024f7900f0ff0447afaa2d18ad6619d5c3e12318R30-R34):
Added a check for deterministic signing by comparing the initial seed
with an additional seed in the `fromEthSigner` function. If they differ,
an error is thrown to indicate the wallet does not support deterministic
signing.
*
[`src/account.ts`](diffhunk://#diff-ae151444b91b7e5e0a6aa753024f7900f0ff0447afaa2d18ad6619d5c3e12318R73-R77):
Implemented a similar check in the `fromStarknetAccount` function by
comparing the initial signature with an additional signature using the
new `isStarknetSignatureDeterministic` function.
*
[`src/account.ts`](diffhunk://#diff-ae151444b91b7e5e0a6aa753024f7900f0ff0447afaa2d18ad6619d5c3e12318R124-R146):
Added the `isStarknetSignatureDeterministic` function to validate the
deterministic nature of Starknet signatures by comparing the components
of the signatures.

Documentation update:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L28):
Removed the unused `[get in touch]` reference link from the
documentation.